+++ submission/persistence-common-api/src/main/kotlin/ac/uk/ebi/biostd/persistence/common/model/SimpleSubmission.kt
