--- submission/persistence-common/src/main/kotlin/ac/uk/ebi/biostd/persistence/common/service/SubmissionMetaQueryService.kt
@@ -1,17 +1,0 @@
-package ac.uk.ebi.biostd.persistence.common.service
-
-import java.io.File
-import java.time.OffsetDateTime
-
-interface SubmissionMetaQueryService {
-    fun getParentAccPattern(accNo: String): String?
-    fun isNew(accNo: String): Boolean
-    fun getSecret(accNo: String): String?
-    fun getAccessTags(accNo: String): List<String>
-    fun getReleaseTime(accNo: String): OffsetDateTime?
-    fun existByAccNo(accNo: String): Boolean
-    fun findCreationTime(accNo: String): OffsetDateTime?
-    fun getCurrentFolder(accNo: String): File?
-    fun getOwner(accNo: String): String?
-    fun isProcessing(accNo: String): Boolean
-}
