+++ submission/persistence/build.gradle.kts
@@ -28,7 +28,7 @@ dependencies {
     api(project(":commons:commons-model-extended"))
     api(project(":commons:commons-model-extended-mapping"))
     api(project(":commons:commons-test"))
-    api(project(":submission:persistence-common"))
+    api(project(":submission:persistence-common-api"))
     api(project(":submission:submission-config"))
 
     implementation(SpringAutoConfigure)
